+++ CHANGELOG.md
@@ -12,6 +12,10 @@
   Getdown. They must be prefixed by `sys.`: for example `sys.silent = true` will set the `silent`
   system property to `true`.
 
+* If Getdown is run in a headless JVM, it will avoid showing a UI but will attempt to install and
+  launch the application anyhow. Note that passing `-Dsilent` will override this behavior (because
+  in silent mode the default is only to install the app, not also launch it).
+
 * Fixed issue with `appid` not being properly used when specified via command line arg.
 
 * Fixed issue with running Getdown on single CPU systems (or virtual systems). It was attempting to